+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - refactored personalization to be more universal and self contained
 - adding step indicators to build and buy configuration app
 - Added Nosto recommendations container divs
+- added additional generic classes
+- added countdown timer for home page
 
 ## [1.20.31]
 - new cart layout and cart functionality